Add button to ChooseTask that submits chosen task
