+++ src/client/components/tasks/TypeButton.tsx
@@ -2,12 +2,11 @@ import React, { useState, useContext } from 'react';
 import axios from 'axios';
 import dayjs from 'dayjs';
 import { Button } from '../../../components/ui/button';
-import { UserContext } from '../../contexts/UserContext';
 
 type TypeButtonProps = {
   type: string;
   taskInfo: TaskInfo;
-  setTask: React.Dispatch<React.SetStateAction<object>>;
+  setTaskInfo: React.Dispatch<React.SetStateAction<TaskInfo>>;
 };
 type TaskInfo = {
   type: string;
@@ -16,35 +15,20 @@ type TaskInfo = {
   userId?: number;
 };
 
-function TypeButton({ type, taskInfo, setTask }: TypeButtonProps) {
-  const { user, getUser } = useContext(UserContext);
-  // Function to assign the chosen task category to the user
-  const pickTask = (element) => {
-    const userId = user.id;
+function TypeButton({ type, taskInfo, setTaskInfo }: TypeButtonProps) {
+  // Function to set the taskInfo type
+  const pickType = (element) => {
     const { id } = element.target;
     const copyTask = { ...taskInfo };
-    const formattedDate = taskInfo.date.format('MM/DD/YYYY');
-    copyTask.date = dayjs(formattedDate);
     copyTask.type = id;
-    const config = {
-      taskInfo: copyTask,
-    };
-    config.taskInfo.userId = userId;
-    // Make axios request: UPDATE the users current task AND create a user_task row
-    axios
-      .post('/api/task', config)
-      .then(({ data }) => {
-        setTask(data);
-      })
-      .then(() => {
-        getUser();
-      })
-      .catch((err) => {
-        console.error('Error posting task: ', err);
-      });
+    setTaskInfo(copyTask);
   };
   return (
-    <Button id={type} onClick={pickTask}>
+    <Button
+      variant={taskInfo.type === type ? 'secondary' : 'outline'}
+      id={type}
+      onClick={pickType}
+    >
       {type}
     </Button>
   );
